fix: move masks.py to generate/
