+++ src/anemoi/graphs/generate/masks.py
