+++ src/anemoi/graphs/nodes/builders/from_file.py
@@ -7,8 +7,8 @@ from anemoi.datasets import open_dataset
 from anemoi.utils.config import DotDict
 from torch_geometric.data import HeteroData
 
+from anemoi.graphs.generate.masks import KNNAreaMaskBuilder
 from anemoi.graphs.nodes.builders.base import BaseNodeBuilder
-from anemoi.graphs.nodes.masks import KNNAreaMaskBuilder
 
 LOGGER = logging.getLogger(__name__)
