+++ src/anemoi/graphs/nodes/builders/from_refined_icosahedron.py
@@ -12,8 +12,8 @@ from torch_geometric.data import HeteroData
 
 from anemoi.graphs.generate.hexagonal import create_hexagonal_nodes
 from anemoi.graphs.generate.icosahedral import create_icosahedral_nodes
+from anemoi.graphs.generate.masks import KNNAreaMaskBuilder
 from anemoi.graphs.nodes.builders.base import BaseNodeBuilder
-from anemoi.graphs.nodes.masks import KNNAreaMaskBuilder
 
 LOGGER = logging.getLogger(__name__)
